requirements.txt
# Production requirements for deployment
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Computer vision
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# ML utilities
scikit-learn>=1.3.0
numpy>=1.24.0
torchmetrics>=1.0.0

# OCR
jiwer>=3.0.0

# API and serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Monitoring and logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Optimization
optuna>=3.4.0

# Model export
onnx>=1.15.0
onnxruntime>=1.16.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0
